fixing swagger
